AIModelsAPI/app/requirements.txt
 # FastAPI Requirements
anyio==3.6.2
click==8.1.3
fastapi
h11==0.14.0
idna==3.7
pydantic
python-multipart==0.0.7
six==1.16.0
sniffio==1.3.0
starlette
uvicorn==0.20.0
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
slowapi~=0.1.7
colorama~=0.4.6
cryptography~=43.0.0
pydantic-settings

# Yolov7 Requirements
pillow~=10.4.0
numpy
requests>=2.23.0
tqdm>=4.41.0
torch>=1.7.0,!=1.12.0
pandas>=1.1.4
opencv-python>=4.1.1
PyYAML>=5.3.1
matplotlib>=3.2.2
seaborn>=0.11.0
scipy>=1.4.1